+++ CHANGELOG.md
@@ -9,11 +9,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## API
 * use IDP3 syntax for 2 theories (environment + decision)
-* add "environmental" of symbols in /meta's response
 * support simple `display` theory, with `goal`, `relevant`, `expand`, `hide`, `view` and `moveSymbols` predicates
   * `view` in /meta can take the `hidden` value
   * replace `__goals(Term)` by `__relevant(Term)`
-* add `propagated` to /meta's response
+* add "environmental" of symbols in /meta's response
+* add "is_assignment' to /eval's response
 
 ## GUI
 * File / New